Add carList, clientList

+++ src/main/java/ru/skipp/autopartstore/AuthUser.java
@@ -12,7 +12,7 @@ public class AuthUser extends org.springframework.security.core.userdetails.User
         this.user = user;
     }
 
-    public int id() {
+    public long id() {
         return user.id();
     }
